requirements.txt
altgraph==0.17.4
packaging==24.1
pefile==2024.8.26
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyshortcuts==1.9.1
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.2